src/turbine/function-deploy/function-app/requirements.txt
aiohttp>=3.8.1
grpcio>=1.44.0
grpcio-tools>=1.44.0
protobuf>=3.20.0
